+++ test/DryIoc.IssuesTests/GHIssue179_MadeOf_Parameters_dont_follow_Reuse_setting.cs
@@ -22,6 +22,38 @@ namespace DryIoc.IssuesTests
             Assert.AreNotSame(b.Log.ContextType, b.A.Log.ContextType);
         }
 
+        [Test]
+        public void Test_with_parameter_selector()
+        {
+            var container = new Container();
+            container.Register<A>();
+            container.Register<B>();
+
+            container.Register(typeof(ILog), typeof(Log),
+                made: Parameters.Of.Type<Type>(r => r.Parent.ImplementationType));
+
+            var b = container.Resolve<B>();
+
+            Assert.AreNotSame(b.Log.ContextType, b.A.Log.ContextType);
+        }
+
+        [Test]
+        public void Test_made_with_parameters()
+        {
+            var container = new Container();
+            container.Register<A>();
+            container.Register<B>();
+
+            container.Register<ILog>(
+                made: Made.Of(
+                    typeof(Log).SingleConstructor(),
+                    parameters: Parameters.Of.Type<Type>(r => r.Parent.ImplementationType)));
+
+            var b = container.Resolve<B>();
+
+            Assert.AreNotSame(b.Log.ContextType, b.A.Log.ContextType);
+        }
+
         public interface ILog
         {
             Type ContextType { get; }